Refactor folder architecture for server
